:metal: add ViewSettingable and fix Rootable
